change AboutCard to SmallCard
