+++ src/components/SmallCard.css
@@ -1,4 +1,4 @@
-.about-card {
+.small-card {
   display: flex;
   flex-wrap: wrap;
   flex-grow: 1;
